Use Rubocop service
Updates `Linter::Ruby` to remove all the in-line linting code and
instead fall-back to the job based approach provided by `Linter::Base`.

We're moving to job names based on linters, so this introduces
`RubocopReviewJob`, which is scheduled by `Linter::Ruby`. I tried
renaming `Linter::Ruby` and `Config::Ruby` to `Rubocop` and the config
key from `ruby` to `rubocop` (with an alias), but so much spec setup and
assumptions rely on the Ruby name that it wasn't worth including as a
part of this change.

https://trello.com/c/jG5McIv8